another llm suggestion

+++ strfmt.h
@@ -37,7 +37,7 @@ const char * vstrfmt(char* buf, size_t bufLen, const char* format, va_list argp)
 #define STRFMT_ADD_CHAR(c) if(bufptr<bufLen-1){buf[bufptr++]=c;}else{buf[bufLen-1]='\0';return buf;}
 const char * vstrfmt(char* buf, const size_t bufLen, const char* format, va_list argp) {
     bool inSpecifier = false;
-    unsigned int bufptr = 0;
+    size_t bufptr = 0;
     if (buf == NULL || bufLen == 0) {
         return NULL;
     }
